[FIX] website: fix footer disappearing when switching theme
When a theme was chosen, the footer disappeared if any footer template
different from the default one was previously selected. The related
code is actually meant to revert back to the default footer template
when switching theme (allowing the theme to set a different default one
if it chooses to), but it forgot to enable that default one... and was
only disabling all non-default footer templates.

closes odoo/odoo#53762

X-original-commit: 7679dfe014b6681ce9ac01c7c45c62fa7a1d6b71
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>