[FIX] hr_timesheet: Prevent the stoping timer when time dialog cancelled
Currently, When some one start the timer and click on stop button at
that time it was first stop the timer by super method and then open
the time dialog so timer vanishes and the Start button is displayed again.

After this commit, when click on stop button do not stop the timer by
calling super method of mixin just open the dialog with minutes spent
and on click of SAVE in dialog stop the time and unlink
the related timer entry and on click of CANCEL of dialog
it will not stop the timer.

closes odoo/odoo#55744

Task-id: 2230217
X-original-commit: ae2c9e8d6a9b34d53c3bf31a3d009b5f9d1eb721
Related: odoo/enterprise#12347
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>