[FIX] mail: make composer use same subtype as message replied to
Before this commit, when clicking on reply in inbox would send a message
whatever the subtype of message the reply is linked to.
This commit makes sure that if the message which we want to reply to
is a note (resp. comment), then the reply will be a note (resp. comment)

task-2282363

closes odoo/odoo#54158

X-original-commit: b3141f8c7bb4a57e63066c5577ef1e108a9c122c
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>